feat: Detect 요청 및 응답 기반 CansView 재랜더링 기능 구현
- 기존 dummy DetectRequest 제거 및 실제 도면 좌표 기반 요청 전송 로직
  반영
- DeteectResponse 수신 시 LiveData에 저장 및 Toast 알림 표시

+++ core/ui/src/main/java/com/binbean/ui/CanvasView.kt
@@ -8,7 +8,6 @@ import android.graphics.Paint
 import android.util.AttributeSet
 import android.util.Log
 import android.view.GestureDetector
-import android.view.LayoutInflater
 import android.view.MotionEvent
 import android.view.ViewGroup
 import android.widget.FrameLayout
@@ -17,7 +16,6 @@ import androidx.core.content.ContextCompat
 import com.binbean.admin.dto.FloorDetail
 import com.binbean.admin.dto.Position
 import com.binbean.domain.cafe.FloorListDto
-import com.binbean.domain.cafe.FloorPlanResponse
 import com.binbean.domain.cafe.PositionDto
 
 class CanvasView @JvmOverloads constructor(
@@ -289,8 +287,8 @@ class CanvasView @JvmOverloads constructor(
         for (i in 0 until childCount) {
             val view = getChildAt(i) as? ImageView ?: continue
             val type = view.tag as? CanvasView.ObjectType ?: continue
-            val x = ((view.x + view.width / 2f) / gridSize).toInt()
-            val y = ((view.y + view.height / 2f) / gridSize).toInt()
+            val x = ((view.x + view.width / 2f) / gridSize)
+            val y = ((view.y + view.height / 2f) / gridSize)
             val pos = Position(x, y)
 
             when (type) {
